configure:  removed all the debug warning flags.  it was impossible             to find real problems when compiling in debug mode.
moved mlame_corr.c into 'misc' directory

removed some 'size_t' and 'unsigned' from get_audio.c

made mp3buffer 'unsigned char*' for the mpglib interface routines.

+++ frontend/get_audio.h
@@ -57,7 +57,7 @@ int lame_decoder(lame_global_flags *gfp,FILE *outf,int skip, char *inPath, char 
 
 
 
-void SwapBytesInWords( short *loc, size_t words );
+void SwapBytesInWords( short *loc, int words );
